requirements.txt
# Generador de Arte con IA - Fractales
# Dependencias principales para cálculo científico y visualización
numpy>=1.24.0
matplotlib>=3.6.0
Pillow>=9.4.0

# Optimización de rendimiento
numba>=0.56.0

# Configuración y utilidades
PyYAML>=6.0
click>=8.1.0

# Opcional para análisis avanzado (futuras extensiones)
scipy>=1.10.0
scikit-image>=0.19.0